Colossus animation change state V1 (broken)
The state change works but the animation does not. More testing is required to determine why it is not working. I believe it is due to the animations having different avatars but I am unsure because I tried switching the avatar and it still did not work. One option is to disable the walking colossus and enable the attacking colossus but it would be better if I could figure out what is messing with the rig